Move angleToRotateBetweenVectors2D to utils
